refactor(folder-structure): group home components under Home folder and update imports
